Made separate file with all types (ContinuationResult, Branch, Event) and helper functions to avoid circular imports.
