+++ src/pycont/__init__.py
@@ -1,3 +1,4 @@
 from .continuation import pseudoArclengthContinuation
+from .plotting import plotBifurcationDiagram
 
-__all__ = ["pseudoArclengthContinuation"]
+__all__ = ["pseudoArclengthContinuation", "plotBifurcationDiagram"]
