+++ src/pycont/plotting.py
@@ -1,0 +1,5 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+def plotBifurcationDiagram():
+    pass
